+++ common/src/main/java/dev/ftb/mods/ftbchunks/client/gui/map/LargeMapScreen.java
@@ -2,7 +2,7 @@ package dev.ftb.mods.ftbchunks.client.gui.map;
 
 import com.mojang.blaze3d.platform.InputConstants;
 import dev.ftb.mods.ftbchunks.FTBChunks;
-import dev.ftb.mods.ftbchunks.client.FTBChunksClient;
+import dev.ftb.mods.ftbchunks.client.FTBChunksKeyMappings;
 import dev.ftb.mods.ftbchunks.client.gui.AddWaypointOverlay;
 import dev.ftb.mods.ftbchunks.client.map.*;
 import dev.ftb.mods.ftbchunks.config.FTBChunksClientConfig;
@@ -207,10 +207,10 @@ public class LargeMapScreen extends BaseScreen {
             return true;
         } else if (key.is(InputConstants.KEY_G) && InputConstants.isKeyDown(Minecraft.getInstance().getWindow(), InputConstants.KEY_F3)) {
             FTBChunksClientConfig.CHUNK_GRID.toggle();
-            FTBChunksClientConfig.saveConfig();
+            FTBChunksClientConfig.saveNeeded();
             dimension.getManager().updateAllRegions(false);
             return true;
-        } else if (FTBChunksClient.openMapKey.matches(key.event())) {
+        } else if (FTBChunksKeyMappings.MAP_KEY.matches(key.event())) {
             closeGui(false);
             return true;
         }