+++ common/src/main/java/dev/ftb/mods/ftbchunks/client/minimap/layers/IconLayerRenderer.java
@@ -8,14 +8,13 @@ import dev.ftb.mods.ftblibrary.math.MathUtils;
 import net.minecraft.client.gui.GuiGraphicsExtractor;
 import net.minecraft.util.Mth;
 import net.minecraft.world.phys.Vec3;
-import org.jetbrains.annotations.UnknownNullability;
 import org.joml.Matrix3x2fStack;
 
 public enum IconLayerRenderer implements MinimapLayerRenderer {
     INSTANCE;
 
     @Override
-    public void extractLayer(@UnknownNullability GuiGraphicsExtractor graphics, Matrix3x2fStack poseStack, MinimapRenderContext ctx) {
+    public void extractLayer(GuiGraphicsExtractor graphics, Matrix3x2fStack poseStack, MinimapRenderContext ctx) {
         for (MapIcon icon : ctx.mapIcons()) {
             // map icons (waypoints, entities...)
             Vec3 pos = icon.getPos(ctx.deltaTracker().getGameTimeDeltaPartialTick(false));